Removed ParsnipCode
