+++ JMG.Parsnip.Library/ParsnipGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using JMG.Parsnip.VSIXProject.CodeWriting;
@@ -6,8 +7,19 @@ using JMG.Parsnip.VSIXProject.SerializedModel;
 
 namespace JMG.Parsnip.VSIXProject
 {
+	/// <summary>
+	/// Parser generator
+	/// </summary>
 	public static class ParsnipGenerator
 	{
+		/// <summary>
+		/// Generates a parser
+		/// </summary>
+		/// <param name="inputString">Specification file</param>
+		/// <param name="outputNamespace">Namespace for the generated class</param>
+		/// <param name="fileName">Specification file name</param>
+		/// <param name="versionString">Version number to print in the header</param>
+		/// <returns>Generated output</returns>
 		public static String Generate(String inputString, String outputNamespace, String fileName, String versionString)
 		{
 			try
@@ -39,7 +51,7 @@ namespace JMG.Parsnip.VSIXProject
 				using (writer.Namespace(outputNamespace))
 				{
 					var interfaceName = $"I{className}RuleFactory";
-					using (writer.Interface(interfaceName, Access.Public))
+					using (writer.Interface(interfaceName, Access.Internal))
 					{
 						foreach (var method in semanticModel.InterfaceMethods)
 						{
@@ -52,7 +64,7 @@ namespace JMG.Parsnip.VSIXProject
 					}
 					writer.EndOfLine();
 
-					using (writer.Class(className, Access.Public))
+					using (writer.Class(className, Access.Internal))
 					{
 						var parseResultClassName = "ParseResult";
 						var parseResultClassNameT = $"{parseResultClassName}<T>";
@@ -86,21 +98,360 @@ namespace JMG.Parsnip.VSIXProject
 							writer.Assign("var result", $"{firstRuleParseMethodName}(input, 0, states, factory)"); // Invocation
 							writer.Return("result?.Node");
 						}
+
+						// Populate dictionary of rule identifiers to method names
+						var ruleMethodNames = semanticModel.Rules.ToDictionary(i => i.RuleIdentifier, i => NameGen.ParseFunctionMethodName(i.RuleIdentifier));
+
+						// Generate method signatures
+						List<Signature> signatures = new List<Signature>();
+						Dictionary<SemanticModel.IParseFunction, Invoker> invokers = new Dictionary<SemanticModel.IParseFunction, Invoker>();
+						foreach (var rule in semanticModel.Rules)
+						{
+							var methodName = ruleMethodNames[rule.RuleIdentifier];
+							var visitor = new GenerateSignaturesVisitor(methodName, isMemoized: true, mustAddSignature: true, ruleMethodNames, signatures, invokers);
+							rule.ParseFunction.ApplyVisitor(visitor, Access.Private);
+						}
 						writer.EndOfLine();
 
-						var parsnipCode = new ParsnipCode();
-						parsnipCode.WriteMethods(writer, interfaceName, semanticModel);
+						// Maybe			
+						var cardParams = new[] {
+							new LocalVarDecl("String", "input"), // Invocation
+							new LocalVarDecl("Int32", "inputPosition"),
+							new LocalVarDecl("PackratState[]", "states"),
+							new LocalVarDecl(interfaceName, "factory"),
+							new LocalVarDecl($"Func<String, Int32, PackratState[], {interfaceName}, ParseResult<T>>", "parseAction") // Invocation
+						};
+						using (writer.Method(Access.Private, true, "ParseResult<T>", "ParseMaybe<T>", cardParams))
+						{
+							writer.VarAssign("result", "parseAction(input, inputPosition, states, factory)"); // Invocation
+							using (writer.If("result != null"))
+							{
+								writer.Return("result");
+							}
+							writer.Return("new ParseResult<T>(default(T), 0)");
+						}
+
+						// Star
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<IReadOnlyList<T>>", "ParseStar<T>", cardParams))
+						{
+							writer.VarAssign("list", "new List<T>()");
+							writer.VarAssign("nextResultInputPosition", "inputPosition");
+							using (writer.While("true"))
+							{
+								writer.VarAssign("nextResult", "parseAction(input, nextResultInputPosition, states, factory)"); // Invocation
+								using (writer.If("nextResult == null"))
+								{
+									writer.SwitchBreak();
+								}
+								writer.LineOfCode("list.Add(nextResult.Node);");
+								writer.Assign("nextResultInputPosition", "nextResultInputPosition + nextResult.Advanced");
+							}
+							writer.Return("new ParseResult<IReadOnlyList<T>>(list, nextResultInputPosition - inputPosition)");
+						}
+
+						// Plus
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<IReadOnlyList<T>>", "ParsePlus<T>", cardParams))
+						{
+							writer.VarAssign("list", "new List<T>()");
+							writer.VarAssign("nextResultInputPosition", "inputPosition");
+
+							writer.VarAssign("firstResult", "parseAction(input, nextResultInputPosition, states, factory)"); // Invocation
+							using (writer.If("firstResult == null"))
+							{
+								writer.Return("null");
+							}
+							writer.LineOfCode("list.Add(firstResult.Node);");
+							writer.Assign("nextResultInputPosition", "nextResultInputPosition + firstResult.Advanced");
+
+							using (writer.While("true"))
+							{
+								writer.VarAssign("nextResult", "parseAction(input, nextResultInputPosition, states, factory)"); // Invocation
+								using (writer.If("nextResult == null"))
+								{
+									writer.SwitchBreak();
+								}
+								writer.LineOfCode("list.Add(nextResult.Node);");
+								writer.Assign("nextResultInputPosition", "nextResultInputPosition + nextResult.Advanced");
+							}
+							writer.Return("new ParseResult<IReadOnlyList<T>>(list, nextResultInputPosition - inputPosition)");
+						}
+						writer.EndOfLine();
+
+						// Delimited
+						var delimParams = new[] {
+							new LocalVarDecl("String", "input"), // Invocation
+							new LocalVarDecl("Int32", "inputPosition"),
+							new LocalVarDecl("PackratState[]", "states"),
+							new LocalVarDecl(interfaceName, "factory"),
+							new LocalVarDecl($"Func<String, Int32, PackratState[], {interfaceName}, ParseResult<T>>", "parseAction"), // Invocation
+							new LocalVarDecl($"Func<String, Int32, PackratState[], {interfaceName}, ParseResult<D>>", "parseDelimiterAction") // Invocation
+						};
+
+						using (writer.Method(Access.Private, true, "ParseResult<IReadOnlyList<T>>", "ParseSeries<T, D>", delimParams))
+						{
+							writer.VarAssign("list", "new List<T>()");
+							writer.VarAssign("nextResultInputPosition", "inputPosition");
+
+							writer.VarAssign("firstResult", "parseAction(input, nextResultInputPosition, states, factory)"); // Invocation
+							using (writer.If("firstResult == null"))
+							{
+								writer.Return("null");
+							}
+							writer.LineOfCode("nextResultInputPosition += firstResult.Advanced;");
+							writer.LineOfCode("list.Add(firstResult.Node);");
+
+							using (writer.While("true"))
+							{
+								writer.VarAssign("delimResult", "parseDelimiterAction(input, nextResultInputPosition, states, factory)"); // Invocation
+								using (writer.If("delimResult == null"))
+								{
+									writer.SwitchBreak();
+								}
+
+								writer.VarAssign("nextResult", "parseAction(input, nextResultInputPosition + delimResult.Advanced, states, factory)"); // Invocation
+								using (writer.If("nextResult == null"))
+								{
+									writer.SwitchBreak();
+								}
+
+								writer.LineOfCode("list.Add(nextResult.Node);");
+								writer.Assign("nextResultInputPosition", "nextResultInputPosition + delimResult.Advanced + nextResult.Advanced");
+							}
+							writer.Return("new ParseResult<IReadOnlyList<T>>(list, nextResultInputPosition - inputPosition)");
+						}
+						//
+						writer.EndOfLine();
+						var typicalParams = new List<LocalVarDecl>
+						{
+							new LocalVarDecl("String", "input"), // Invocation
+							new LocalVarDecl("Int32", "inputPosition"),
+							new LocalVarDecl("PackratState[]", "states"),
+							new LocalVarDecl(interfaceName, "factory")
+						};
+						// Any Character
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_AnyCharacter", typicalParams))
+						{
+							using (writer.If("inputPosition >= input.Length"))
+							{
+								writer.Return("null");
+							}
+							writer.Return("new ParseResult<String>(input.Substring(inputPosition, 1), 1)");
+						}
+
+						// Any Letter
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_AnyLetter", typicalParams))
+						{
+							using (writer.If("inputPosition >= input.Length"))
+							{
+								writer.Return("null");
+							}
+							using (writer.ElseIf("!Char.IsLetter(input[inputPosition])"))
+							{
+								writer.Return("null");
+
+							}
+							writer.Return("new ParseResult<String>(input.Substring(inputPosition, 1), 1)");
+						}
+
+
+						// Any Digit
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_AnyDigit", typicalParams))
+						{
+							using (writer.If("inputPosition >= input.Length"))
+							{
+								writer.Return("null");
+							}
+							using (writer.ElseIf("!Char.IsDigit(input[inputPosition])"))
+							{
+								writer.Return("null");
+
+							}
+							writer.Return("new ParseResult<String>(input.Substring(inputPosition, 1), 1)");
+						}
+
+						// End of Line
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_EndOfLine", typicalParams))
+						{
+							writer.VarAssign("result1", "ParseLexeme(input, inputPosition, \"\\r\\n\")"); // Invocation
+							using (writer.If("result1 != null"))
+							{
+								writer.Return("new ParseResult<String>(result1.Node, result1.Advanced)");
+							}
+
+							writer.VarAssign("result2", "ParseLexeme(input, inputPosition, \"\\n\")"); // Invocation
+							using (writer.If("result2 != null"))
+							{
+								writer.Return("new ParseResult<String>(result2.Node, result2.Advanced)");
+							}
+
+							writer.Return("null");
+						}
+
+						// End of Stream
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<EmptyNode>", "ParseIntrinsic_EndOfStream", typicalParams))
+						{
+							using (writer.If("inputPosition == input.Length"))
+							{
+								writer.Return("new ParseResult<EmptyNode>(EmptyNode.Instance, 0)");
+							}
+							writer.Return("null");
+						}
+
+						// End of Line Or Stream
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<EmptyNode>", "ParseIntrinsic_EndOfLineOrStream", typicalParams))
+						{
+							using (writer.If("inputPosition == input.Length"))
+							{
+								writer.Return("new ParseResult<EmptyNode>(EmptyNode.Instance, 0)");
+							}
+
+							writer.VarAssign("result1", "ParseLexeme(input, inputPosition, \"\\r\\n\")"); // Invocation
+							using (writer.If("result1 != null"))
+							{
+								writer.Return("new ParseResult<EmptyNode>(EmptyNode.Instance, result1.Advanced)");
+							}
+
+							writer.VarAssign("result2", "ParseLexeme(input, inputPosition, \"\\n\")"); // Invocation
+							using (writer.If("result2 != null"))
+							{
+								writer.Return("new ParseResult<EmptyNode>(EmptyNode.Instance, result2.Advanced)");
+							}
+
+							writer.Return("null");
+						}
+
+						// C-String
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_CString", typicalParams))
+						{
+							writer.VarAssign("resultStart", "ParseLexeme(input, inputPosition, \"\\\"\")"); // Invocation
+							writer.IfNullReturnNull("resultStart");
+							writer.VarAssign("nextInputPosition", "inputPosition + resultStart.Advanced");
+							writer.VarAssign("sb", "new System.Text.StringBuilder()");
+							using (writer.While("true"))
+							{
+								writer.VarAssign("inputPosition2", "nextInputPosition");
+								writer.VarAssign("resultEscape", "ParseLexeme(input, inputPosition2, \"\\\\\")"); // Invocation
+								using (writer.If("resultEscape != null"))
+								{
+									writer.Assign("inputPosition2", "inputPosition2 + resultEscape.Advanced");
+									writer.VarAssign("resultToken", "ParseIntrinsic_AnyCharacter(input, inputPosition2, states, factory)"); // Invocation
+									writer.IfNullReturnNull("resultToken");
+									using (writer.Switch("resultToken.Node"))
+									{
+										writer.LineOfCode("case \"\\\\\":");
+										writer.LineOfCode("case \"\\\"\":");
+										writer.LineOfCode("case \"t\":");
+										writer.LineOfCode("case \"r\":");
+										writer.LineOfCode("case \"n\":");
+										using (writer.BraceScope())
+										{
+											writer.LineOfCode("sb.Append(\"\\\\\");");
+											writer.LineOfCode("sb.Append(resultToken.Node);");
+											writer.SwitchBreak();
+										}
+										using (writer.SwitchDefault())
+										{
+											writer.Return("null");
+										}
+									}
+									writer.Assign("inputPosition2", "inputPosition2 + resultToken.Advanced");
+
+									writer.Continue();
+								}
+								writer.VarAssign("resultEnd", "ParseLexeme(input, inputPosition2, \"\\\"\")"); // Invocation
+								using (writer.If("resultEnd != null"))
+								{
+									writer.Return("new ParseResult<String>(sb.ToString(), inputPosition2 + resultEnd.Advanced - inputPosition)");
+								}
+								writer.VarAssign("resultChar", "ParseIntrinsic_AnyCharacter(input, inputPosition2, states, factory)"); // Invocation
+								writer.IfNullReturnNull("resultChar");
+								writer.LineOfCode("sb.Append(resultChar.Node);");
+								writer.Assign("nextInputPosition", "inputPosition2 + resultChar.Advanced");
+							}
+						}
+
+						// Optional Horizontal Whitespace
+						writer.EndOfLine();
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseIntrinsic_OptionalHorizontalWhitespace", typicalParams))
+						{
+							using (writer.If("inputPosition >= input.Length"))
+							{
+								writer.Return("new ParseResult<String>(String.Empty, 0)");
+							}
+							writer.VarAssign("sb", "new System.Text.StringBuilder()");
+							writer.VarAssign("nextInputPosition", "inputPosition");
+							using (writer.While("nextInputPosition < input.Length"))
+							{
+								writer.VarAssign("ch", "input[nextInputPosition]");
+								using (writer.If("ch != \' \' && ch != \'\\t\'"))
+								{
+									writer.LineOfCode("break;");
+								}
+								writer.LineOfCode("nextInputPosition++;");
+							}
+							writer.Return("new ParseResult<String>(input.Substring(inputPosition, nextInputPosition - inputPosition), nextInputPosition - inputPosition)");
+						}
+						writer.EndOfLine();
+
+						// Write Lexeme
+						using (writer.Method(Access.Private, true, "ParseResult<String>", "ParseLexeme", new[] {
+							new LocalVarDecl("String", "input"), // Invocation
+							new LocalVarDecl("Int32", "inputPosition"),
+							new LocalVarDecl("String", "lexeme") }))
+						{
+							writer.VarAssign("lexemeLength", "lexeme.Length");
+							writer.IfTrueReturnNull("inputPosition + lexemeLength > input.Length");
+							writer.IfTrueReturnNull("input.Substring(inputPosition, lexemeLength) != lexeme");
+							writer.Return("new ParseResult<String>(lexeme, lexemeLength)");
+						}
+
+						// Generate methods
+						var generateMethods = new GenerateMethodsVisitor(writer, interfaceName, invokers);
+						var commentVisitor = new CommentParseFunctionVisitor(showHeader: true);
+						foreach (var methodItem in signatures)
+						{
+							writer.EndOfLine();
+
+							var target = methodItem.Func;
+							writer.Comment(target.ApplyVisitor(commentVisitor));
+							var returnType = target.ReturnType.GetParseResultTypeString();
+							var name = methodItem.Name;
+
+							using (writer.Method(methodItem.Access, true, returnType, name, typicalParams))
+							{
+								if (methodItem.IsMemoized)
+								{
+									var memName = NameGen.MemoizedFieldName(name);
+									writer.LineOfCode($"if (states[inputPosition].{memName} is var mem && mem != null) {{ return mem; }}");
+									writer.EndOfLine();
+								}
+
+								target.ApplyVisitor(generateMethods, methodItem);
+							}
+						}
 						writer.EndOfLine();
 
 						var packratStateClassName = "PackratState";
 						using (writer.Class(packratStateClassName, Access.Private))
 						{
-							using (writer.Constructor(Access.Public, packratStateClassName, new LocalVarDecl[] { }))
+							foreach (var methodItem in signatures)
 							{
-							}
-							writer.EndOfLine();
+								var returnType = methodItem.Func.ReturnType.GetParseResultTypeString();
+								var memName = NameGen.MemoizedFieldName(methodItem.Name);
 
-							parsnipCode.WriteMemoizedFields(writer);
+								if (methodItem.IsMemoized)
+								{
+									writer.MemberVariable(memName, returnType, Access.Internal);
+								}
+							}
 						}
 					}
 				}
