+++ JMG.Parsnip.Library/SyntacticModel/IParsnipDefinitionItem.cs
@@ -7,21 +7,21 @@ using JMG.Parsnip.VSIXProject.Visiting;
 
 namespace JMG.Parsnip.SyntacticModel
 {
-	public interface IParsnipDefinitionItem
+	internal interface IParsnipDefinitionItem
 	{
 		void ApplyVisitor(IParsnipDefinitionItemVisitor visitor);
 
 		TOutput ApplyVisitor<TOutput>(IParsnipDefinitionItemFuncVisitor<TOutput> visitor);
 	}
 
-	public interface IParsnipDefinitionItemVisitor :
+	internal interface IParsnipDefinitionItemVisitor :
 		IActionVisitor<Rule>,
 		IActionVisitor<Comment>
 	{
 
 	}
 
-	public interface IParsnipDefinitionItemFuncVisitor<TOutput> :
+	internal interface IParsnipDefinitionItemFuncVisitor<TOutput> :
 		IFuncVisitor<Rule, TOutput>,
 		IFuncVisitor<Comment, TOutput>
 	{
